Velocidades em graus/s e area alcancavel desenhada em polares
Duas queixas com a mesma raiz e uma de desenho.

UM BRACO MAIS RAPIDO QUE O OUTRO
Hz significa coisas diferentes em cada junta. Com a engrenagem do
operador -- reducao 16,5 na junta 1 e 4 na junta 2 -- os mesmos 3000 Hz
davam 6,5 graus/s numa e 27 na outra: 4,1 vezes de diferenca, sem ajuste
que igualasse sem refazer a conta a mao. A rampa tinha o mesmo problema:
17 graus/s2 numa e 72 na outra.
- velNormal, velPrecisao, velAuto passam a ser GRAUS/s; a rampa de cada
  junta passa a ser GRAUS/s2. Cada eixo converte com o seu proprio
  passosPorGrau e FREQ_PULSO_MAX_HZ segue sendo o teto do driver
- moverCoordenado passa a receber graus/s: quem manda no tempo e a junta
  com mais GRAUS a percorrer, nao com mais passos
- chaves NOVAS no NVS (velNg/velPg/velAg/acel1g/acel2g): reler 3000 Hz
  como 3000 graus/s seria absurdo, entao quem atualiza recebe os padroes
- o painel mostra quantos Hz cada junta vai pedir ao driver, que e onde
  se ve se algum eixo esta perto do teto do T3D

PERDA DE PASSO
- rampa igual entre as juntas remove uma causa provavel (o eixo de menor
  reducao acelerava 4x mais)
- jogAtualizar so reprograma setSpeedInHz e runForward quando o valor
  muda de fato; antes refazia a rampa a cada milissegundo, o que suja o
  trem de pulsos
- LEIA-ME ganhou a ordem de investigacao: rampa, buffer 5 V (familia HCT,
  nao HC) e ligar os fios de ALM

AREA ALCANCAVEL SO MOSTRANDO UM PEDACO
O desenho tracava a borda do retangulo de limites das juntas e
preenchia. A cinematica direta e 2-para-1, entao essa borda se cruza
sozinha e o preenchimento sai com buracos -- o "yin-yang" da foto.
- redesenhada em coordenadas polares: cada valor de theta2 da UM raio e
  theta1 varre um arco nesse raio, entao a regiao sai solida e correta
- espessura de faixa local e supressao de arcos redundantes, senao a
  borda externa acumulava alfa e virava um aro escuro grosso

TESTES
- G01 com a engrenagem real do operador: as duas juntas percorrem o
  mesmo angulo por segundo, chegam juntas num movimento coordenado e tem
  a mesma rampa angular
- A15 e E/F passaram a esperar a ETAPA do assistente em vez de tempo
  fixo: com velAuto em graus/s o retorno ao zero tem outra duracao
- expectativas de A05c, A15h e A16 convertidas para a unidade nova
- firmware 72 -> 75 verificacoes; interface 46 / 0

+++ RoboCNC/estado.h
@@ -44,7 +44,7 @@ struct Junta {
   // correcao de resolucao.
   float grausHome     = 0.0f;
 
-  uint32_t aceleracao = ACEL_PADRAO;
+  float    aceleracao = ACEL_PADRAO;   // graus/s2
   bool     alarme     = false;
 
   // Sentido do eixo. Se a fiacao do DIR estiver invertida em relacao ao
@@ -63,10 +63,12 @@ extern Junta J2;
 // ---------------------------------------------------------------------
 // Parametros de configuracao (persistidos em NVS)
 // ---------------------------------------------------------------------
-extern uint32_t velNormal;
-extern uint32_t velPrecisao;
-extern uint32_t velAuto;
-extern float    velCordaoMmS;
+// Em graus por segundo. Ver a nota em config.h: Hz nao servia porque
+// significa velocidades diferentes em juntas com engrenagens diferentes.
+extern float velNormal;
+extern float velPrecisao;
+extern float velAuto;
+extern float velCordaoMmS;   // mm/s: o cordao ja e especificado assim
 
 extern float elo1Mm;
 extern float elo2Mm;
@@ -179,9 +181,9 @@ void definirMensagem(const char* fmt, ...);
 // para as variaveis vivas num ponto seguro do ciclo.
 // ---------------------------------------------------------------------
 struct ConfigPendente {
-  uint32_t velNormal, velPrecisao, velAuto;
-  float    velCordaoMmS;
-  uint32_t acel1, acel2;
+  float    velNormal, velPrecisao, velAuto;   // graus/s
+  float    velCordaoMmS;                      // mm/s
+  float    acel1, acel2;                      // graus/s2
   uint32_t ppv1, ppv2;
   float    red1, red2;
   bool     inv1, inv2;
